test(cloud-gate): split the escape behavior file under the LOC ceiling
Codex round 20 (P2): the behavior file had grown to 489 lines, past the
repo's ~400 ceiling. Extract the shared fixture helpers and move the
wired gate integration cases into their own file, so the denylist unit
cases and the gate chain stay independently maintainable.
